# bot-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Habit Tracker Bot",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/ping": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "handler"
                ],
                "summary": "Ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/activateHabit": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "handler"
                ],
                "summary": "Activate user habit",
                "parameters": [
                    {
                        "description": "HabitActivationRequest",
                        "name": "habitActivation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.HabitActivationModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "handler"
                ],
                "summary": "Create bot user",
                "parameters": [
                    {
                        "description": "CreateUserRequest",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UserModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "handler"
                ],
                "summary": "Get bot user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid formatted ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UserModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.HabitActivationModel": {
            "type": "object",
            "properties": {
                "active_habit_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "user_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "handler.UserModel": {
            "type": "object",
            "properties": {
                "active_habit_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "birthday": {
                    "type": "string",
                    "format": "datetime"
                },
                "createdAt": {
                    "type": "string",
                    "format": "datetime"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "nickname": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}